Lots of little changes to libretroshare. Improvements mainly focused on configuration storage and loading:  * Improved Configuration Manager (almost finished)  * Mutex protections for Configuration system  * added Configuration storage to p3ConnectMgr.  * added Configuration storage to p3MsgService.  * bugfixes in p3GeneralConfig.  * Added Config Save notification where necessary.  * added safe FinalConfigSave before exit().  * added RsPeerNetItem + RsPeerStunItem (serialiser)  * reordered startup for correct config loading.  * enabled Loading of certs in AuthXPGP.  * move sockaddr_clear() to util/rsnet.h  * switched p3MsgService sendMessage checking to pqiMonitor syste,.  * corrected pqiarchive saving of PeerIds.  * added setNetworkMode() & setVisState() to p3ConnectMgr.  * added Mutex protection to p3ranking.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@336 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/pqi/pqiindic.h
@@ -36,15 +36,15 @@
 class Indicator
 {
 	public:
-	Indicator(int n = 1)
+	Indicator(uint16_t n = 1)
 	:num(n), changeFlags(n) {IndicateChanged();}
 void	IndicateChanged()
 	{
-		for(int i = 0; i < num; i++)
+		for(uint16_t i = 0; i < num; i++)
 			changeFlags[i]=true;
 	}
 
-bool	Changed(int idx = 0)
+bool	Changed(uint16_t idx = 0)
 	{
 		/* catch overflow */
 		if (idx > num - 1)
@@ -56,7 +56,7 @@ bool	Changed(int idx = 0)
 	}
 
 	private:
-	int num;
+	uint16_t num;
 	std::vector<bool> changeFlags;
 };
